moved delete if no dependant entries to data access layer to use a single db pool connection

+++ l_DataAccess/tipo_documento.js
@@ -96,17 +96,35 @@ module.exports = {
         })
     },
 
-    deleteEntry : (id) => {
-        pool.getConnection((err, connection) => {
-            if(err) throw err
-            connection.query('DELETE FROM tipoDocumento WHERE idTipo = ?', id, (err, rows) => {
-                connection.release()
-                if (!err) {
-                    console.log(`Tipo de documento con id ${id} ha sido eliminado`)
-                } else {
-                    console.log(err)
-                }
+    deleteEntry : (id, callback) => {
+        const dataPromise = new Promise((resolve, reject) => {
+            pool.getConnection((err, connection) => {
+                if(err) throw err
+                connection.query('select * from documento where tipo = ?', id, (err, rows) => {
+                    if (!err) {
+                        if(rows.length != 0){
+                            connection.release()
+                            resolve(false)
+                        }
+                        else {
+                            connection.query('DELETE FROM tipoDocumento WHERE idTipo = ?', id, (err, rows) => {
+                                connection.release()
+                                if (!err) {
+                                    console.log(`Tipo de documento con id ${id} ha sido eliminado`)
+                                    resolve(true)
+                                } else {
+                                    reject(console.log(err))
+                                }
+                            })
+                        }
+                    } else {
+                        reject(console.log(err))
+                    }
+                })
             })
+        })
+        dataPromise.then(isDeleted => {
+            callback(isDeleted)
         })
     }
 }
